requirements.txt
<<<<<<< HEAD
torch
transformers
openai-whisper
ffmpeg-python
opencv-python
pytesseract
faiss-cpu
fastapi
uvicorn
=======
torch
transformers
openai-whisper
ffmpeg-python
opencv-python
pytesseract
faiss-cpu
fastapi
uvicorn
>>>>>>> 1f657a2 (Add audio extraction and speech-to-text modules)
